requirements.txt
requests
python-dotenv
tqdm
loguru
